// src/types/api.ts
export interface AuthUser {
  id: string
  email: string
  role: 'admin' | 'customer'
  name: string
}

export interface LoginRequest {
  phone_or_email: string
  password: string
}

export interface LoginResponse {
  user: AuthUser
  token: string
  expiresIn: string
}

export interface RegisterRequest {
  email: string
  password: string
  name: string
  phone: string
  address?: string
  address_type?: 'apartment' | 'rumah' | 'company'
}